Scripts_ETL/requirements.txt
google-cloud-bigquery
pandas
openpyxl
psycopg2-binary